requirements.txt
streamlit
pandas
numpy
plotly
google-generativeai
chardet
openpyxl
xlsxwriter
python-dateutil
pytz
scikit-learn
statsmodels
pytest